+++ src/library/cli/R/cliapp-docs.R
@@ -82,7 +82,8 @@
 #' * `run` is an R expression, that is potentially clickable if the terminal
 #'   supports ANSI hyperlinks to runnable code (e.g. RStudio).
 #'   It supports link text. See [links] for more information about cli hyperlinks.
-#' * `str` for a double quoted string escaped by [base::encodeString()].#' * `strong` for strong importance.
+#' * `str` for a double quoted string escaped by [base::encodeString()].
+#' * `strong` for strong importance.
 #' * `topic` is a help page of a _topic_.
 #'   If the terminal supports ANSI hyperlinks to help pages (e.g. RStudio),
 #'   then cli creates a clickable link. It supports link text.
@@ -543,7 +544,7 @@ NULL
 #' convert relative paths, and paths starting with `~` to aboslute path.
 #'
 #' ```{asciicast links-file-1}
-#' cli_text("... edit your {.file ~/.Rprofile} file.}")
+#' cli_text("... edit your {.file ~/.Rprofile} file.")
 #' ```
 #'
 #' ## Link text