+++ src/library/cli/R/num-ansi-colors.R
@@ -309,10 +309,11 @@ emacs_version <- function() {
   ver <- Sys.getenv("INSIDE_EMACS")
   if (ver == "") return(NA_integer_)
 
-  ver <- gsub("'", "", ver)
+  ver <- gsub("'", "", ver, fixed = TRUE)
+
   ver <- strsplit(ver, ",", fixed = TRUE)[[1]]
   ver <- strsplit(ver, ".", fixed = TRUE)[[1]]
-  as.numeric(ver)
+  suppressWarnings(as.numeric(ver))
 }
 
 win10_build <- function() {