+++ src/library/pkgdepends/R/find-package-root.R
@@ -1,4 +1,11 @@
+project_root_anchors <-
+  c("DESCRIPTION", ".git", ".Rproj.user", "renv.lock", "renv")
+
 find_package_root <- function(path = ".") {
+  find_project_root(path = path, anchors = "DESCRIPTION")
+}
+
+find_project_root <- function(path = ".", anchors = project_root_anchors) {
   is_root <- function(path) {
     identical(
       normalizePath(path, winslash = "/"),
@@ -17,7 +24,7 @@ find_package_root <- function(path = ".") {
   )
   max_depth <- 100
   for (i in 1:max_depth) {
-    if (file.exists(file.path(cur_path, "DESCRIPTION"))) {
+    if (any(file.exists(file.path(cur_path, anchors)))) {
       return(cur_path)
     } else if (is_root(cur_path)) {
       stop(errmsg)