+++ src/library/pkgdepends/R/utils.R
@@ -7,6 +7,8 @@ pkgd_data <- new.env(parent = emptyenv())
 
 `%||%` <- function(l, r) if (is.null(l)) r else l
 
+`%|0|%` <- function(l, r) if (length(l) == 0) r else l
+
 `%|z|%` <- function(l, r) if (is.null(l) || identical(l, "")) r else l
 
 `%&z&%` <- function(l, r) if (length(l) > 0 && l != "") r else ""
@@ -502,4 +504,13 @@ collapse <- function(x, ...) {
 
 na_omit <- function(x) {
   x[!is.na(x)]
+}
+
+file_ext <- function(x) {
+  re_match(x, "[.]([[:alnum:]]+)$")[[".match"]]
+}
+
+# drop a prefix and a postfix, vectorized
+omit_pre_post <- function(x, pre = 0, post = 0) {
+  substr(x, 1L + pre, nchar(x) - post)
 }