+++ src/library/pkgcache/NEWS.md
@@ -1,5 +1,7 @@
 # pkgcache (development version)
 
+# pkgcache 2.2.3
+
 * The metadata cache now does not use source URLs for packages in `Archive`
   on Posit Package Manager repositories. This URLs may serve a different
   package, even a source package when the main URL for the same package